requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1

PyPDF2==3.0.1

nltk==3.9.2
scikit-learn==1.5.2

sumy==0.11.0

pywebview==4.4.1

python-dotenv==1.0.1
